% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MazamaCovidUtils.R
\docType{data}
\name{ArchiveBaseDir}
\alias{ArchiveBaseDir}
\title{Base directory for pre-generated data}
\format{
Directory string.
}
\description{
If an archive of pre-generated data files is available locally,
users can set the location of this directory with\code{setArchiveBaseDir()}.
Otherwise, users must specify an external source of pre-generated datafiles 
with \code{setArchiveBaseUrl()}.

To avoid internet latency, specification of BASE_DIR will always take 
precedence over specification of BASE_URL.

Package functions that load pre-generated data files will load data from this
directory.
}
\seealso{
\code{\link{ArchiveBaseDir}}, \code{\link{getArchiveBaseDir}},  
\code{\link{setArchiveBaseDir}}, \code{\link{removeArchiveBaseDir}}

\code{\link{ArchiveBaseUrl}}, \code{\link{getArchiveBaseUrl}},  
\code{\link{setArchiveBaseUrl}}, \code{\link{removeArchiveBaseUrl}}
}
\keyword{environment}
