% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MazamaCovidUtils.R
\name{getArchiveBaseDir}
\alias{getArchiveBaseDir}
\title{Get data archive base directory}
\usage{
getArchiveBaseDir()
}
\value{
directory string.
}
\description{
Returns the package base directory pointing to an archive of
pre-generated data files.
}
\seealso{
\code{\link{ArchiveBaseDir}}, \code{\link{getArchiveBaseDir}},  
\code{\link{setArchiveBaseDir}}, \code{\link{removeArchiveBaseDir}}

\code{\link{ArchiveBaseUrl}}, \code{\link{getArchiveBaseUrl}},  
\code{\link{setArchiveBaseUrl}}, \code{\link{removeArchiveBaseUrl}}
}
\keyword{environment}
