% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MazamaCovidUtils.R
\docType{package}
\name{MazamaCovidUtils}
\alias{MazamaCovidUtils}
\title{Data access and analysis functions for COVID-19 related data}
\description{
This package contains code to access current COVID-19 data from
various sources.
}
