% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MazamaCovidUtils.R
\name{removeArchiveBaseUrl}
\alias{removeArchiveBaseUrl}
\title{Remove data archive base URL}
\usage{
removeArchiveBaseUrl()
}
\value{
Silently returns previous value of the base URL.
}
\description{
Resets the data archive base URL to NULL. Used for internal 
testing.
}
\seealso{
\code{\link{ArchiveBaseDir}}, \code{\link{getArchiveBaseDir}},  
\code{\link{setArchiveBaseDir}}, \code{\link{removeArchiveBaseDir}}

\code{\link{ArchiveBaseUrl}}, \code{\link{getArchiveBaseUrl}},  
\code{\link{setArchiveBaseUrl}}, \code{\link{removeArchiveBaseUrl}}
}
\keyword{environment}
