% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MazamaCovidUtils.R
\name{setArchiveBaseDir}
\alias{setArchiveBaseDir}
\title{Set data archive base directory}
\usage{
setArchiveBaseDir(archiveBaseDir)
}
\arguments{
\item{archiveBaseDir}{Base directory pointing to an archive of pre-generated 
data files.}
}
\value{
Silently returns previous value of base directory.
}
\description{
Sets the package base directory pointing to an archive of
pre-generated data files.
}
\seealso{
\code{\link{ArchiveBaseDir}}, \code{\link{getArchiveBaseDir}},  
\code{\link{setArchiveBaseDir}}, \code{\link{removeArchiveBaseDir}}

\code{\link{ArchiveBaseUrl}}, \code{\link{getArchiveBaseUrl}},  
\code{\link{setArchiveBaseUrl}}, \code{\link{removeArchiveBaseUrl}}
}
\keyword{environment}
